Fixed one issue on the no data message and added the cap to words for text inputs.  Also deleted a table from the database that was no longer needed, checking before trying to fix the match lists grouping.

+++ My NRA Rifle Match Score Sheet/LIST_Matches.m
@@ -179,7 +179,7 @@
             iRow++;
         }
     }
-    [FormFunctions doBuggermeMessage:[NSString stringWithFormat:@"%@ has %ld rows",displayMatchesClasses.matchclass,iRow] FromSubFunction:@"list_matches.getRowsForSection"];
+    [FormFunctions doBuggermeMessage:[NSString stringWithFormat:@"%@ has %ld rows",displayMatchesClasses.matchclass,(long)iRow] FromSubFunction:@"list_matches.getRowsForSection"];
     return iRow;
 }
